breaking-feat: complete Clone Hierarchy feat: Hybrid and MultiHybrid validation feat: Hybrid and MultiHybrid Tests
This Commit introduces the new Hierarchy used.
An Abstract class Hybrid is used for MultiHybrid and Hybrid which defines all necessary methods.
Additionally, Methods are provided to create either Concrete IV or IC Objects (like the factory pattern).
While I would not call it great Design in terms of Software Engineering,
I think it is not that bad in this use case, because the only reason IV and IC Clones
exist is in order to improve the query efficiency since it divides the graph for e.g "MultiHybrid" into "ICMultiHybrid" and "IVMultiHybrid" and makes selection easier. The factory pattern could be introduced at a later point, but it would increase the amount of classes and therefore was not implemented for now.