refactor: Change Assertion to Custom Exception remove: Lavel does not check for existence of Nepenthes name anymore
Before this commit label, there was a list that was filled in to check the existence of a type before creation. However, I think the existence should be checked in the service, and therefore the model no longer checks for it. Another reason for that is also the added complexity and coupling between the Subclasses: A hybrid contains two species (e.g villosa x hamata) which must be checked, however certains hybrid are known under a new name e.g xRokko. This means in this case we would need to check for species, but also for the hybridname in PrimaryHybrid. This is then also true for Multihybrids