refactor: change Nepenthes to Species refactor and add: NepenthesManagementApi refactor: NepenthesManagement Controller per Tag feat: add custom Query to return label via labelClass,CloneType and nepenthesName
This Commit refactored mostly the Controller and the OpenAPI since these changes are necessary for the (big) upcoming Angular application extension and refactoring commit.
At a Later date the Controller classes will most likely be generated in one big Controller Class to save Resources.
Apart from this "Nepenthes" was also changed to "Species" with the intention to use from now on "Nepenthes" as an abstract construct (e.g. villosa is a Nepenthes, but villosa x hamata a Primary hybrid is also a Nepenthes)