remove and refactor: Growlistmanagement uses NepenthesManagementOpenAPI.yaml instead of an own openAPi specification. refactor: minor changes in OpenAPI and Controllers for Growlistmanagement feat: adding existing Nepenthes (clones) and new Nepenthes (clones) is working in Growlistmanagement add: Integration test for Growlistmanagement add: new custom Exceptions add: proxy Method in NepenthesManagementService in order to create new nepenthes using only String instead of Label.
This Commit introduces functionality in order to add new and existing clones of a Nepenthes to an User.
Furthermore, some changes were made. Especially the openAPI Specification for NepenthesManagement and Growlist were merged, since Growlist can be seen as an extension for NepenthesManagement. However, it will be still an own Component since Growlist depends on UserManagement and therefore would introduce a dependency for NepenthesManagement which is not wanted (to depend on User)