add: DTO for the GrowlistManagement which were missing
Due to the missing DTO the workflow failed. for now a bit of duplication is introduced since the DTO are the same as in the NepenthesManagement.
In the future however, Nepflow is supposed to be several microservices and therefore at some point the dependency would have to be removed anyway