fix and performance: retrieving and storing new Specimen to the Growlist add: custom Queries for Specimen prepare: ImageService for storing the images into the Filesystem add: testcase for ImageService refactor: Growlist Testcases
The custom Queries which did retrieve the growlist and all its Specimen.
However, storing and saving a new Specimen in the Growlist did delete the reference from clone to label.
To fix this issue and also increase the performance custom Queries were introduced which add a Specimen without touching the Growlist.

ImageService was introduced to be able to later save Images into the Filesystem. However, before using this approach I would like to try Spring Content first.

Instead, on depending on a files to insert specific information for the Growlist test, LabelCloneDefinitions was introduced.
Some refactoring will be necessary in the Future.