refactor: Unit Tests feat: add PollenExchange Component
This Commit refactored the Unit Tests. Instead of a static Class which defines models, a DataInserter was created which uses the Repository to insert the data.  This DataInserter is extended depending on the "dependency" of  another Module. e.g. the Growlist needs Labels and Clones in order to test inserting specimens. The PollenExchange needs Labels and Clones as well, but also specimens in order to test creating PollenOffers and Trades.