refactor: Specimen has reference to user refactor: PollenOffer has no reference to User anymore add: unit tests feat: add several precondition into Trade and PollenOffer refactor: replace logic into domain model
This Update is mostly improved the domain model  (rich domain model vs anemic).
E.g. the PollenOffer now knows it´s own expiry Date and  therefore the service simply needs to call the PollenOffer to check if it is still usable.

Another Example is the Trade which now provides a Method in order to check if the given User is allowed to accept or refuse the Trade. The Service itself must only call the Method without needing to know the internal  details  of the Trade.

the added preconditions are for example to  make  sure  that  a PollenOffer is only initialised if  the Specimen itself  has a Sex, because in this Domain it would  make little sense to  offer a flower/pollen  if the sex is unknown.
Another example is the Trade which now makes sure, that both Offers have a different User or are both valid/open.