king-feat: Refactor Project to Standalone Component breaking-feat: Change from Css to SASS refactor: change Project structure add: tests
This Commit introduces the new Project structure and the upgrade to
Standalone Components. Furthermore, the Chance was taken to also switch
to SASS. The CSS However was not yet changed to SASS which means some
site may look worse than normally.
At last ng test is now also executable and runs without any error and
will later be added to the workflow(github actions)breaking-feat: