fix: remove bug from last commit message refactor: decompose  PollenOverview component designer choice: pollen-overview has global variable for  child component
The PollenOverview component was decomposed in:
- basecard component which renders one mat card  and also allows to add further content using ng-content
- nepenthes-card-dropdown extends the basecard image which also fixes the bug since every dropdown is now an own component
About the designer-choice:
- The PollenOverview still has the lookup and fills it. It could be easily done without for better separation of concern, but due to performance reason I won´t remove it for now. The lookup is needed to in order to find out which user PollenOffer ist already connected to another PollenOffer in order to prevent trade requests where both  PollenOffers were already referenced.
Without global lookup every dropdown would need check the trades of the user, the pollenOffers of the User and the PollenOffer of the other User. This would result in way more loops and comparison and therefore was not implemented like this.

the problem itself is that the global structure (lookup) is used and modified in the dropdown components.