refactor: introduce enum SEX for class Sex (#16) fix: user-first-step sends request to backend again remove: dead code in user-growlist.component.ts feat: enable change of unknown sex in growlist overview (close #9)
This Commit refactored the Sex class.
There is now an Enum "SEX" which contains the valid values for Sex which are either "MALE" or "FEMALE". In addition, this enum also contains "UNKNOWN" which is used if the sex itself is null. It is not used as a default because this would mean:
1. an extreme amount of unnecessary references since in order to find clones without sex simply checking for non existence of male and female is possible.
2. having null instead of unknown simplifies the adding of plants, because by default clones without sex have the property cloneId = internalCloneId