dependency: com.mashape.unirest for requests dependency: io.jsonwebtoken:jjwt for jwt checking feat: Auth0UserServiceImpl  (#20) test: add tests for Auth0UserServiceImpl (#20) refactor: add new properties for Auth0UserServiceImpl (#20) refactor: workflow using github secret to enable decrypting (#20)
This commit introduces the feature (#20) and some tests.
About the tests: While it is better to mock the external systems, I decided to not do it in this case.
The reason is that the service itself relies on the answers and the format of Auth0. Meaning if Auth0 does change endpoint/format, the tests would still succeed while in production the service would fail.
Therefore, no mocking is done, and I intend to only execute these tests locally and in milestones to prevent failed workflows due to rate limits or downtime.