feat: ChatModule (#12) dependency: postgressql, spring-boot-starter-data-jpa, javax.xml.bind add: UnitTests for ModelMapper and ChatModule add: Gradle Task for openAPi for ChatModule refactor: UserCreatedEvent now uses primitive types instead of User object (introduces a dependency (UserRetrievalService) in order to be able to reference the actual User Object) (#16) update: application.yaml
This commit introduces the new ChatModule, which will be utilized in future features. The module uses an RDBMS for storage. Although Neo4j was a viable option, the decision was made to use an RDBMS due to its compatibility with future deployment strategies, particularly the intent to eventually separate the module in terms of deployment.
( Additionally, an RDBMS fits better for a chat application than a graph database, as it aligns more naturally with the data structure and relationships typical of a chat application.)